scripts/requirements.txt
# Additional requirements for scripts
requests>=2.31.0
python-dotenv>=1.0.0
